pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "consistency"
version = "0.1.0"
description = "Semantic alignment scoring for text-to-image consistency"
requires-python = ">=3.8"
dependencies = [
 "torch>=1.13",
 "transformers>=4.30",
 "Pillow",
 "numpy",
 "tqdm",
 "opencv-python",
 "ultralytics",
]

[project.scripts]
consistency = "consistency.cli:main"
